Add files via upload
This project is a GUI-based Simple Calculator implemented using Python and the tkinter library. This calculator allows users to perform basic arithmetic operations (addition, subtraction, division and multiplication) on two numbers. Users can input the numbers into the entry fields and select the operation by clicking on the dropdown list of operation buttons (+, -, /,*), and then click the "Calculate" button to obtain the output. The calculated result is displayed in a separate entry field. The calculator GUI is a user-friendly layout for ease of use.