Add files via upload
This Rock-Paper-Scissor game project is a simple graphical user interface (GUI) application created using Python and the tkinter library. It allows users to play the vintage Rock-Paper-Scissor game against a computer(CodSoft) opponent. The game features an attractive and user-friendly interface with colorful buttons and labels is provided in this project.